refactor: refactor client side
