+++ client/src/api/slide.ts
@@ -16,31 +16,23 @@ const slideInstance = axios.create({
   headers: {authorization: localStorage.getItem("token")},
 });
 
-export const slide = {
-  get({ id }: Props): Promise<AxiosResponse<Props>> {
+const slideApi = {
+  get({id}: Props): Promise<AxiosResponse<Props>> {
     return slideInstance.get(`${id}`);
   },
   getAll(): Promise<AxiosResponse<Array<Props>>> {
     return slideInstance.get("/");
   },
-  create({ data }: Props): Promise<AxiosResponse<Props>> {
+  create({data}: Props): Promise<AxiosResponse<Props>> {
     return slideInstance.post("/", data);
   },
-  update({ id, data }: Props): Promise<AxiosResponse<Props>> {
+  update({id, data}: Props): Promise<AxiosResponse<Props>> {
     return slideInstance.put(`${id}`, data);
   },
-  delete({ id }: Props): Promise<AxiosResponse<Props>> {
+  delete({id}: Props): Promise<AxiosResponse<Props>> {
     return slideInstance.delete(`${id}`);
   },
 };
 
-const slidesInstance = axios.create({
-  baseURL: "localhost:8080/api/search",
-});
-
-export const slides = {
-  getAll(): Promise<AxiosResponse<Array<Props>>> {
-    return slidesInstance.get("/");
-  },
-};
+export default slideApi;
