+++ client/src/assets/icons/index.ts
@@ -1,0 +1,20 @@
+import logo from "./logo.svg";
+import menu from "./menu.svg";
+import user from "./user.svg";
+import recent from "./recent.svg";
+import slide from "./slide.svg";
+import archive from "./archive.svg";
+import newSlides from "./newSlides.svg";
+import more from "./more.svg";
+
+export {
+  logo,
+  menu,
+  user,
+  recent,
+  slide,
+  archive,
+  newSlides,
+  more,
+};
+
