+++ client/src/components/common/Cards.tsx
@@ -1,16 +1,54 @@
 import React from "react";
+import styled from "@emotion/styled";
 import Card from "./Card";
-import * as S from "./assets";
+import {MOBILE_MAX_WIDTH} from "../../domains/constants";
+
+const CardsLayout = styled.div`
+  padding-bottom: 3rem;
+
+  > h2 {
+    color: #fff;
+    margin: 0;
+    display: inline-block;
+    font-size: 2rem;
+  }
+  
+  > hr {
+    border-bottom: 1px solid #777;
+    width: 7rem;
+    margin: 1rem 0 2rem;
+  }
+`;
+
+const CardsBlock = styled.div`
+  display: grid;
+  grid-gap: 20px;
+  grid-template-columns: repeat(4, 1fr);
+  
+  a {
+    text-decoration: none;
+    color: #000;  
+  }
+  
+  @media (max-width: 1050px) {
+    grid-template-columns: repeat(3, 1fr);
+  }
+  
+  @media (max-width: ${MOBILE_MAX_WIDTH}px) {
+    grid-template-columns: repeat(2, 1fr);
+  }
+`;
+
 
 interface IProps {
   title: string;
 }
 
 const Cards: React.FC<IProps> = ({title}) => (
-  <S.CardsLayout>
+  <CardsLayout>
     <h2>{title}</h2>
     <hr/>
-    <S.Cards>
+    <CardsBlock>
       <Card id={1} title="Hello World!" subtitle="Introducing Limetree" author="Hodol" createdAt="2020-07-12"/>
       <Card id={1} title="Hello World!" subtitle="Introducing Limetree" author="Hodol" createdAt="2020-07-12"/>
       <Card id={1} title="Hello World!" subtitle="Introducing Limetree" author="Hodol" createdAt="2020-07-12"/>
@@ -18,8 +56,8 @@ const Cards: React.FC<IProps> = ({title}) => (
       <Card id={1} title="Hello World!" subtitle="Introducing Limetree" author="Hodol" createdAt="2020-07-12"/>
       <Card id={1} title="Hello World!" subtitle="Introducing Limetree" author="Hodol" createdAt="2020-07-12"/>
       <Card id={1} title="Hello World!" subtitle="Introducing Limetree" author="Hodol" createdAt="2020-07-12"/>
-    </S.Cards>
-  </S.CardsLayout>
+    </CardsBlock>
+  </CardsLayout>
 );
 
 export default Cards;
