+++ client/src/components/common/SidebarHeader.tsx
@@ -1,5 +1,46 @@
 import React from "react";
-import * as S from "./assets";
+import styled from "@emotion/styled";
+import * as S from "../../assets/icons";
+
+const SidebarHeaderBlock = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  width: 100%;
+  padding-bottom: 1rem;
+  border-bottom: 2px solid #aaa;
+  margin-bottom: 1rem;
+  
+  > div {
+    font-size: 1.4rem;
+    display: flex;
+    align-items: center;
+    flex: 5;
+    
+    > img {
+      width: 1.4rem;
+      padding-right: 10px;
+    }
+  }
+  
+  > button {
+    border: none;
+    background: transparent;
+    flex: 1;
+    cursor: pointer;
+    display: flex;
+    align-items: center;
+    justify-content: flex-end;
+    
+    &:focus {
+      outline: none;
+    }
+    
+    > img {
+      width: 0.75rem;
+    } 
+  }
+`;
 
 interface IProps {
   src?: string;
@@ -7,14 +48,14 @@ interface IProps {
 }
 
 const SidebarHeader: React.FC<IProps> = ({src, title}) => (
-  <S.SidebarHeader>
+  <SidebarHeaderBlock>
     <div>
       <img src={src || S.user} alt={title}/><span>{title}</span>
     </div>
     <button>
       <img src={S.more} alt="more"/>
     </button>
-  </S.SidebarHeader>
+  </SidebarHeaderBlock>
 );
 
 export default SidebarHeader;
