+++ client/src/components/editor/Preview.tsx
@@ -1,15 +1,23 @@
 import React from "react";
-import * as S from "./assets";
-import Markdown from "../common/markdown";
+import styled from "@emotion/styled";
+import Markdown from "../markdown";
 
 interface IProps {
   content: string
 }
 
+const PreviewBlock = styled.div`
+  background-color: #fff;
+  flex: 1;
+  padding: 30px;
+  overflow: scroll;
+  max-height: 100vh;
+`;
+
 const Preview: React.FC<IProps> = ({content}) => (
-  <S.Preview>
+  <PreviewBlock>
     <Markdown value={content}/>
-  </S.Preview>
+  </PreviewBlock>
 );
 
 export default Preview;
