+++ client/src/components/markdown/renderers/Image.tsx
