+++ client/src/domains/constants.ts
@@ -1,6 +1,7 @@
-// export const PRIMARY_GREEN = "#76BA1B";
-export const PRIMARY_GREEN = "#1E1E1E";
+export const ZEZE_GRAY = "#1E1E1E";
+export const PRIMARY_GREEN = "#76BA1B";
 export const SECONDARY_GREEN = "#58C247";
+
 export const MAX_WIDTH = 960;
 export const MOBILE_MAX_WIDTH = 575.98;
