--- client/src/domains/types.ts
@@ -1,7 +1,0 @@
-export interface Markdown {
-  content: string;
-  delimiter: string;
-}
-
-export type Language = string | undefined;
-export type Replacer = (code: string) => string;
