--- client/src/hooks/index.ts
@@ -1,2 +1,0 @@
-// export {default as useInput} from "./useInput";
-export {};
