+++ client/src/index.tsx
@@ -1,11 +1,29 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import "./index.css";
+import {css, Global} from "@emotion/core";
 import App from "./App";
 import * as serviceWorker from "./serviceWorker";
 
+const globalStyle = css`
+  body {
+    background-color: #222;
+    margin: 0;
+    font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
+    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
+    sans-serif;
+    -webkit-font-smoothing: antialiased;
+    -moz-osx-font-smoothing: grayscale;
+  }
+  
+  code {
+    font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
+    monospace;
+  }
+`;
+
 ReactDOM.render(
   <React.StrictMode>
+    <Global styles={globalStyle}/>
     <App/>
   </React.StrictMode>,
   document.getElementById("root"),
