+++ client/src/pages/Me.tsx
@@ -1,7 +1,7 @@
 import React from "react";
-import SidebarLayout from "../common/SidebarLayout";
-import Cards from "../common/Cards";
-import Toast from "../common/Toast";
+import SidebarLayout from "../components/common/SidebarLayout";
+import Cards from "../components/common/Cards";
+import Toast from "../components/common/Toast";
 
 const Me: React.FC = () => (
   <SidebarLayout>
