--- client/src/utils/index.ts
@@ -1,2 +1,0 @@
-// export {default as DateUtil} from "./DateUtil";
-export {};
